# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BileMo",
        "description": "Serveur BileMo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API BileMo"
        }
    ],
    "paths": {
        "/login_check": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Envoi un JWT token",
                "operationId": "9056e654130817d5830fe3912b502e81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "username du client",
                                        "type": "string",
                                        "example": "votrePseudo"
                                    },
                                    "password": {
                                        "description": "Password du client",
                                        "type": "string",
                                        "format": "password",
                                        "example": "votrePassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Liste des produits",
                "operationId": "66ae5815a7406e5f3fea8e33147a7c14",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Le numéro de la page à récupérer",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Détail d'un produit",
                "operationId": "1cf77d4e079e193e161f2e4fca2e53a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail d'un produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Liste des users",
                "operationId": "c7e4897e26d4d8d62b2e3d8c6bf3e53c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Le numéro de la page à récupérer",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un user",
                "operationId": "f16746d02f8019ff08b5449dee239ead",
                "requestBody": {
                    "request": "AddUser",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creation user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Détail d'un user",
                "operationId": "750fe8d308ad6fa2328bf19c00004af6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail d'un user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Effacer un user",
                "operationId": "8b476306d78082d28f66738db9b1b843",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete User"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "JWT not found or expired"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "json"
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "date_creation": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "date_update": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "json"
                    },
                    "password": {
                        "description": "The hashed password",
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "date_creation": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "date_update": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}